requirements.txt
absl-py==0.11.0
astor==0.8.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
gast==0.4.0
grpcio==1.35.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.1.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
mock==3.0.5
numpy==1.16.2
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.15.1
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
urllib3==1.26.3
Werkzeug==1.0.1
zipp==1.2.0
